video: refactor and extract client code into /modules
