+++ apps/web/app/future/video/[uid]/page.tsx
@@ -4,8 +4,9 @@ import { WithLayout } from "app/layoutHOC";
 
 import { APP_NAME } from "@calcom/lib/constants";
 
+import type { PageProps } from "~/videos/views/videos-single-view";
 import VideosSingleView from "~/videos/views/videos-single-view";
-import { getServerSideProps, type PageProps } from "~/videos/views/videos-single-view.getServerSideProps";
+import { getServerSideProps } from "~/videos/views/videos-single-view.getServerSideProps";
 
 export const generateMetadata = async () =>
   await _generateMetadata(
